+++ TaskExplorer/API/Windows/SandboxieAPI.h
@@ -2,6 +2,10 @@
 
 #include <QObject>
 
+#define CONF_GET_NO_GLOBAL          0x40000000L
+#define CONF_GET_NO_EXPAND          0x20000000L
+#define CONF_GET_NO_TEMPLS          0x10000000L
+
 class CSandboxieAPI : public QObject
 {
 	Q_OBJECT
@@ -25,7 +29,13 @@ public:
 
 	void QueryPathList(quint64 ProcessId, quint32 path_code, QStringList& Paths) const;
 
-	QList<QPair<QString, QString>> GetIniSection(const QString& BoxName, qint32* pStatus = NULL, bool withTemplates = true) const;
+	struct SbieIniValue
+	{
+		QString Name;
+		quint32 Type = 0;
+		QString Value;
+	};
+	virtual QList<SbieIniValue> GetIniSection(const QString& BoxName, qint32* pStatus = NULL, bool withTemplates = true, bool withGlobals = true) const;
 
 	quint32 QueryProcessInfoEx(quint64 ProcessId, quint32* pil = NULL, quint32* pit = NULL);