+++ TaskExplorer/API/Windows/WinAdmin.cpp
@@ -320,9 +320,9 @@ bool SkipUacEnable (bool is_enable)
 	return result;
 }
 
-bool SkipUacRun (bool test_only)
+int SkipUacRun(bool test_only)
 {
-	bool result = false;
+	int result = 0;
 
 	ITaskService* service = nullptr;
 	ITaskFolder* folder = nullptr;
@@ -337,7 +337,7 @@ bool SkipUacRun (bool test_only)
 
 	wchar_t szPath[MAX_PATH];
 	if (!GetModuleFileName(NULL, szPath, ARRAYSIZE(szPath)))
-		return false;
+		return result;
 
 	MBSTR root (L"\\");
 	MBSTR name (SKIP_UAC_TASK_NAME);
@@ -375,7 +375,7 @@ bool SkipUacRun (bool test_only)
 											{
 												if (test_only)
 												{
-													result = true;
+													result = 1;
 												}
 												else
 												{
@@ -400,12 +400,11 @@ bool SkipUacRun (bool test_only)
 													if (SUCCEEDED(registered_task->RunEx(params, TASK_RUN_NO_FLAGS, 0, nullptr, &running_task)))
 													{
 														UINT8 count = 3; // try count
+														TASK_STATE state = TASK_STATE_UNKNOWN;
 
 														do
 														{
 															QThread::msleep(250);
-
-															TASK_STATE state = TASK_STATE_UNKNOWN;
 
 															running_task->Refresh();
 															running_task->get_State(&state);
@@ -421,12 +420,15 @@ bool SkipUacRun (bool test_only)
 																	state == TASK_STATE_READY
 																	)
 																{
-																	result = true;
+																	result = 1;
 																}
 
 																break;
 															}
 														} while (count--);
+
+														if(state == TASK_STATE_UNKNOWN)
+															result = -1;
 
 														running_task->Release();
 													}