+++ TaskExplorer/GUI/TaskExplorer.h
@@ -11,7 +11,7 @@
 
 #define VERSION_MJR		1
 #define VERSION_MIN 	6
-#define VERSION_REV 	0
+#define VERSION_REV 	1
 #define VERSION_UPD 	0
 
 class CGraphBar;